examples/components/theme_changer/requirements.txt
streamlit-component-theme-changer